added services

+++ src/main/java/com/auth/auth/response/UserInfoResponse.java
@@ -3,19 +3,19 @@ package com.auth.auth.response;
 import java.util.List;
 
 public class UserInfoResponse {
-    private Long id;
+    private int id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public UserInfoResponse(Long id, String username, String email, List<String> roles) {
+    public UserInfoResponse(int id, String username, String email, List<String> roles) {
         this.id = id;
         this.username = username;
         this.email = email;
         this.roles = roles;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
@@ -32,7 +32,7 @@ public class UserInfoResponse {
     }
 
     
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
